+++ tsconfig.json
@@ -8,7 +8,8 @@
     "jsx": "react-jsx",
     "moduleResolution": "node",
     "resolveJsonModule": true,
-    "target": "ES2019",
+    "target": "ES2020",
+    "module": "esnext",
     "strict": true,
     "allowJs": true,
     "forceConsistentCasingInFileNames": true,
